+++ dom-tags/class-picture-tag.php
@@ -1,37 +1,34 @@
 <?php
 /**
- * The <source> DOMtag.
- * @since 1.2.0
+ * The <picture> DOMtag.
+ * @since 1.2.4
  *
  * @author Jace Fincham
  * @package DomTags
  */
 namespace DomTags;
 
-class SourceTag extends \DomTags implements DomTagInterface {
+class PictureTag extends \DomTags implements DomTagInterface {
 	/**
 	 * Construct the DOMtag.
-	 * @since 1.2.0
+	 * @since 1.2.4
 	 *
 	 * @access public
 	 * @param array|null $args (optional) -- The list of arguments.
 	 * @return string
 	 */
 	public static function tag(?array $args = null): string {
-		return parent::constructTag('source', self::props(), $args);
+		return parent::constructTag('picture', self::props(), $args);
 	}
 	
 	/**
 	 * The tag's props.
-	 * @since 1.2.0
+	 * @since 1.2.4
 	 *
 	 * @access public
 	 * @return array
 	 */
 	public static function props(): array {
-		return array_merge(
-			parent::ALWAYS_WL,
-			array('src', 'srcset', 'type', 'media')
-		);
+		return parent::ALWAYS_WL;
 	}
 }